fix: antiasan.c

+++ lab5/antiasan.c
@@ -2,7 +2,7 @@
 void antiasan(unsigned long addr)
 {
     // find shadow 
-    unsigned* shadow = (unsigned*) ( 0x7fff8000 + (addr >> 3));
+    unsigned long* shadow = (unsigned long*) ( 0x7fff8000 + (addr >> 3));
     // destroy the shadow;
     *shadow = 0;
 }
